Open /dev/urandom instead of /dev/random since we have no proper entropy source yet.
Fix annoying data abort - the crash was in a function called __udivdi3,
which is apparently provided by the toolchain. The 'smaller' of the two
mp_div functions in libtommath does not get linked with __udivdi3,
and does not crash. So use it instead of the 'faster' one.
We should probably fix __udivdi3 in our toolchain in the long term.


git-svn-id: http://svn.dslinux.org/svn/trunk@1187 ee40c8e4-d140-46bc-8bdd-8ae3c7fe0c09