Initial shot at adding support for ext2 to the RAM build.
This is not like GBAMP-EXT2, i.e. the root filesystem will
remain a romfs filesystem. The ext2 filesystem is mounted on /media.

I've added ext2 support to the kernel config, and changed
/boot/rc to take ext2 filesystems on the second primary
partition into account.

The annoying tab-completion issue I've been having on my 2GB SD card
is now gone, so apparently Amadeus was right and it's been indeed
caused by a bug in the FAT filesystem code.

I've renamed the /dev/mmcblk0 and /dev/mmcblk1 devices
to /dev/mmc and /dev/mmc1 to make scripting filesystem mount in
/boot/rc a little easier. And I've added /dev/mmc2 to be able to
mount the second partition.

One issue that still remains is that we are currently unable
to properly unmount the ext2 filesystem when the system is powered off,
so the filesystem needs to be fsck'd before it can be used again.
I've added umount to busybox but failed to unmount /media
manually before powering off because the filesystem was busy.
I will look into that next.


git-svn-id: http://svn.dslinux.org/svn/trunk@1699 ee40c8e4-d140-46bc-8bdd-8ae3c7fe0c09