Disable sed again.
The build keeps jumping into subdirectories that fail to build.
For example, docs/ - building docs fails because help2man tries
to run the cross-compiled sed binary on the host system.

Trying to narrow down the list of subdirectories to build
in user/sed/Makefile fails, because that file is sometimes
regenerated for some reason.

So I am disabling this until someone fixes this port so
it builds reliably. Preferrably make this port use its
configure script properly please.


git-svn-id: http://svn.dslinux.org/svn/trunk@1827 ee40c8e4-d140-46bc-8bdd-8ae3c7fe0c09