Apply telnetd patch by Sonny Jim:
I've just checked out a clean svn for the first time in a while and I
noticed that telnetd still doesn't work on the DLDI build.  Here's a
patch that fixes it to run through inetd which works fine.  Also it
creates some entries for /etc/services, which it didn't before



git-svn-id: http://svn.dslinux.org/svn/trunk@1880 ee40c8e4-d140-46bc-8bdd-8ae3c7fe0c09